Progress bar with seeking, and some other nice fixes.
remove loop around SDL_PutAudio*.  Previously we had a fixed size buffer
where we didn't know how much we coudl take from the current frame vs.
how much we could fit in the intermediate buffer vs. how much additional
bytes the sdl audiostream wanted.  this meant we needed more than one
iteration, but with SDL_PutAudio we can just immediately put
a. the entire frame
b. additional_amoun bytes
whichever is smaller.  Because the codec may give us planar data, we
need at least, sdl 3.4.0, so that we can use
SDL_PutAudioStreamPlanarData in that case.

Show currently playing track name at the bottom of the window.

Pass filename as parameter to avformat_open_input instead of allocating
our own AVIOContext with a mapped file.  For some reason, having our own
custom AVIOContext makes avformat_find_stream_info not find out about
the duration of the track.

Set avcodec log to quiet.  We otherwise get messages about the duration
being estimated from the bitrate, which may be inaccurate.  We don't
care.

Show progress bar when a track is playing.  The relative progress
requires some estimation, by using the current frame from the decoder
and its timestamp, but it looks ok.

The progress bar can also be clicked to seek to a position.

Also fix a race condition when accessing current frame or packet.  The
audio callback runs in a different thread from the main thread.  So
whenever we try to load audio and kill the current frame or packet, and
the audio callback thread is trying to use that frame or packet, that's
a race condition.  Now there's a lock whenever we access frame or
packet.  Technically there's also a race condition when we read the
last_relative_duration field which is set in the audio callback thread
but read in the draw_progress_bar function, but reading stale data there
shouldn't be a problem.  We'd only draw a progress bar that's slightly
out of date.