shuffle and autoplay, keys X,S,N,B,G,Q
shuffle is missing a history. so we can't use N,B to move backwards and
forwards in the shuffle history.  and also we can't force an even
distribution that doesn't have random repeats.

also fix broken eof handling that would just continue to ask for packets
at the end of the stream.